Part 3 - Find team with smallest goal difference

+++ src/main/scala/part2/Triangle.scala
@@ -1,9 +1,9 @@
 package part2
 
-case class Triangle(base: Double, height: Double) {
+case class Triangle(base: BigDecimal, height: BigDecimal) {
   require(base != 0)
   require(height !=0)
-  def area: Double = {
+  def area: BigDecimal = {
     (base * height) / 2
   }
 }
